Move common voice agent functionality into base class
