+++ backend/service/agents/voice_agent.py
@@ -6,16 +6,11 @@ from langgraph.store.base import BaseStore
 from langfuse.langchain import CallbackHandler
 
 from service.utils.environment import REDIS_HOST
-from service.utils.wav_utils import load_audio_from_file
-from service.model import HuggingFaceGemma3nClient
+from service.agents.voice_agent_base import VoiceAgentBase
 from service.prompts import COMMUNICATION_AGENT_PROMPT
 
 
-class VoiceCommunicationAgent:
-    def __init__(self):
-        model_obj = HuggingFaceGemma3nClient()
-        self.model, self.processor = model_obj.model, model_obj.processor
-
+class VoiceCommunicationAgent(VoiceAgentBase):
     def __call_model(
         self, state: MessagesState, config: RunnableConfig, *, store: BaseStore
     ):
@@ -26,42 +21,12 @@ class VoiceCommunicationAgent:
         info = "\n".join([str(d.value) for d in memories])
 
         system_msg = COMMUNICATION_AGENT_PROMPT.format(info=info)
-
-        messages = [
-            {
-                "role": "user",
-                "content": [
-                    {
-                        "type": "audio",
-                        "audio": load_audio_from_file(
-                            file_path=state["messages"][-1].content
-                        ),
-                    },
-                ],
-            },
-            {
-                "role": "system",
-                "content": [{"type": "text", "text": system_msg}],
-            },
-        ]
-
-        inputs = self.processor.apply_chat_template(
-            messages,
-            add_generation_prompt=True,
-            tokenize=True,
-            return_dict=True,
-            return_tensors="pt",
-        ).to(self.model.device, dtype=self.model.dtype)
-
-        with torch.inference_mode():
-            out = self.model.generate(
-                **inputs, max_new_tokens=256, disable_compile=True
-            )
-
-        response = self.processor.decode(
-            out[0][inputs["input_ids"].shape[-1] :], skip_special_tokens=True
+        audio_path = state["messages"][-1].content
+        messages = self.construct_model_messages(
+            audio_path=audio_path, system_msg=system_msg
         )
-        return {"messages": response}
+
+        return {"messages": self.model.invoke(messages)}
 
     def generate(self, prompt: str):
         with RedisStore.from_conn_string(REDIS_HOST) as store:
