+++ backend/service/agents/voice_memory_agent.py
@@ -10,57 +10,23 @@ import logging
 from pathlib import Path
 
 from service.utils.environment import REDIS_HOST
-from service.utils.wav_utils import load_audio_from_file
-from service.model import HuggingFaceGemma3nClient
+from service.agents.voice_agent_base import VoiceAgentBase
 from service.prompts import MEMORY_EXTRACTION_PROMPT
 
 
 logger = logging.getLogger(__name__)
 
 
-class VoiceMemoryAgent:
-    def __init__(self):
-        model_obj = HuggingFaceGemma3nClient()
-        self.model, self.processor = model_obj.model, model_obj.processor
-
+class VoiceMemoryAgent(VoiceAgentBase):
     def __store_memory(
         self, state: MessagesState, config: RunnableConfig, *, store: BaseStore
     ):
-        messages = [
-            {
-                "role": "user",
-                "content": [
-                    {
-                        "type": "audio",
-                        "audio": load_audio_from_file(
-                            file_path=state["messages"][-1].content
-                        ),
-                    },
-                ],
-            },
-            {
-                "role": "system",
-                "content": [{"type": "text", "text": MEMORY_EXTRACTION_PROMPT}],
-            },
-        ]
-
-        inputs = self.processor.apply_chat_template(
-            messages,
-            add_generation_prompt=True,
-            tokenize=True,
-            return_dict=True,
-            return_tensors="pt",
-        ).to(self.model.device, dtype=self.model.dtype)
-
-        with torch.inference_mode():
-            out = self.model.generate(
-                **inputs, max_new_tokens=256, disable_compile=True
-            )
-
-        response = self.processor.decode(
-            out[0][inputs["input_ids"].shape[-1] :], skip_special_tokens=True
+        audio_path = state["messages"][-1].content
+        messages = self.construct_model_messages(
+            audio_path=audio_path, system_msg=MEMORY_EXTRACTION_PROMPT
         )
-        return {"messages": response}
+
+        return {"messages": self.model.invoke(messages)}
 
     def __get_memory_to_store(
         self, state: MessagesState, config: RunnableConfig, *, store: BaseStore
