+++ backend/service/model/hf_client.py
@@ -21,3 +21,23 @@ class HuggingFaceGemma3nClient:
         )
         self.processor = AutoProcessor.from_pretrained(self.MODEL_ID)
         logging.info("Loaded HuggingFace Gemma3n model.")
+
+    def invoke(self, messages):
+        inputs = self.processor.apply_chat_template(
+            messages,
+            add_generation_prompt=True,
+            tokenize=True,
+            return_dict=True,
+            return_tensors="pt",
+        ).to(self.model.device, dtype=self.model.dtype)
+
+        with torch.inference_mode():
+            out = self.model.generate(
+                **inputs, max_new_tokens=256, disable_compile=True
+            )
+
+        response = self.processor.decode(
+            out[0][inputs["input_ids"].shape[-1] :], skip_special_tokens=True
+        )
+
+        return response
